+++ zaifbot/modules/api/order.py
@@ -5,8 +5,8 @@ from threading import Thread, Event
 from uuid import uuid4
 from zaifbot.bot_common.errors import ZaifBotError
 from zaifbot.bot_common.logger import logger
-from zaifbot.bot_common.utils import get_current_last_price
 from zaifbot.modules.api.cache import ZaifCurrencyPairs
+from zaifbot.modules.utils import get_current_last_price
 
 
 class AutoCancelClient:
